Move files about in beacon sim
