+++ tests/beacon_chain_sim/src/local_network.rs
